[refactor][m]: Rename name of package to match name of repo
