+++ frictionless_ckan_mapper/ckan_to_frictionless.py
