+++ frictionless_ckan_mapper/converter.py
