+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-package=ckan_datapackage_tools
+package=frictionless_ckan_mapper
 skip_missing_interpreters=true
 envlist=
   py36
